Implemented SetResistance for new RealBike with corresponding bluetooth methods

+++ RemoteHealthcare/bike/SimulatorBike.cs
@@ -33,7 +33,7 @@ namespace RemoteHealthcare.bike
             services.GetService<DeviceManager>().HandleData(data);
         }
 
-        public void SetResistance(int resistance)
+        public void SetResistance(byte resistance)
         {
             throw new NotImplementedException();
         }
